extract choice card stories into separate modules
