+++ src/core/components/choice-card/stories/single-state-icon-mobile.tsx
@@ -1,0 +1,44 @@
+import React from "react"
+import { ThemeProvider } from "emotion-theming"
+import { storybookBackgrounds } from "@guardian/src-helpers"
+import { SvgCamera, SvgAudio, SvgVideo } from "@guardian/src-icons"
+import { ChoiceCardGroup, ChoiceCard, choiceCardDefault } from "../index"
+
+/* eslint-disable react/jsx-key */
+const iconChoiceCards = [
+	<ChoiceCard
+		value="picture"
+		label="Picture"
+		id="picture"
+		icon={<SvgCamera />}
+	/>,
+	<ChoiceCard
+		value="audio"
+		label="Audio"
+		id="audio"
+		icon={<SvgAudio />}
+		checked={true}
+	/>,
+	<ChoiceCard value="video" label="Video" id="video" icon={<SvgVideo />} />,
+]
+/* eslint-enable react/jsx-key */
+
+export const singleStateWithIconMobileLight = () => (
+	<ThemeProvider theme={choiceCardDefault}>
+		<ChoiceCardGroup name="colours" label="Media format">
+			{iconChoiceCards.map((choiceCard, index) =>
+				React.cloneElement(choiceCard, { key: index }),
+			)}
+		</ChoiceCardGroup>
+	</ThemeProvider>
+)
+
+singleStateWithIconMobileLight.story = {
+	name: `single state with icon mobile light`,
+	parameters: {
+		backgrounds: [
+			Object.assign({}, { default: true }, storybookBackgrounds.default),
+		],
+		viewport: { defaultViewport: "mobileMedium" },
+	},
+}
