+++ src/core/components/choice-card/stories/single-state-mobile.tsx
@@ -1,0 +1,37 @@
+import React from "react"
+import { ThemeProvider } from "emotion-theming"
+import { storybookBackgrounds } from "@guardian/src-helpers"
+import { ChoiceCardGroup, ChoiceCard, choiceCardDefault } from "../index"
+
+/* eslint-disable react/jsx-key */
+const choiceCards = [
+	<ChoiceCard value="red" label="Red" id="default-red" />,
+	<ChoiceCard
+		value="green"
+		label="Green"
+		id="default-green"
+		checked={true}
+	/>,
+	<ChoiceCard value="blue" label="Blue" id="default-blue" />,
+]
+/* eslint-enable react/jsx-key */
+
+export const singleStateMobileLight = () => (
+	<ThemeProvider theme={choiceCardDefault}>
+		<ChoiceCardGroup name="colours" label="What is your favourite colour?">
+			{choiceCards.map((choiceCard, index) =>
+				React.cloneElement(choiceCard, { key: index }),
+			)}
+		</ChoiceCardGroup>
+	</ThemeProvider>
+)
+
+singleStateMobileLight.story = {
+	name: `single state mobile light`,
+	parameters: {
+		backgrounds: [
+			Object.assign({}, { default: true }, storybookBackgrounds.default),
+		],
+		viewport: { defaultViewport: "mobileMedium" },
+	},
+}
